front: drop eslint-config-airbnb
This package's latest release is from 2021 (4 years ago), and it's
not compatible with ESLint v9. Moreover it contains a lot of
irrelevant (because we use TypeScript instead of plain JavaScript)
or highly questionable (var declarations on top, no loops, and so
on) rules.

Replace it with recommended rule sets from upstream packages. This
also brings osrd's ESLint config closer to osrd-ui's.

Signed-off-by: Simon Ser <contact@emersion.fr>