front: move path properties state into usePathfinding()
Right now, the path properties state is defined outside of
usePathfinding(), and the setter is passed around. This makes it
tricky to figure out the data flow: the setter can be called from
anywhere. Additionally, this makes it tempting to overwrite the
state outside of usePathfinding().

Move the state inside usePathfinding() and don't expose the setter.
Return the state so that a read-only view of the state is provided.
This ensures implementation details are kept inside usePathfinding().

Signed-off-by: Simon Ser <contact@emersion.fr>