gateway: fix proxying of responses without a body
Some HTTP responses do not and cannot have a body, which
is most http clients translate as an empty body. If we
attempt to stream an empty body with a response code which
cannot have a body, actix does not raise an error and fails
to add the transfer-encoding header, but still transmits
the content, which is not HTTP compliant.

Do not attempt to stream the body when the backend response
does not have one.

Signed-off-by: Victor "multun" Collod <victor.collod@epita.fr>