front: adds opId into PathOperationalPoint
This commit aims at clarifying the OperationalPoint types (there are two
identical types) that in practice have IDs that are not actual OP IDs,
because the Manchette needs unique IDs, basically.

The solution is to also store the actual OP ID when it's possible, and
get types with a slightly better naming, PathOperationalPoint, to
highlight that these represent operational points, within the context of
a given path.

Details:
- Replaces type OperationalPoint from both operationalStudies/types and
  stdcm/types, by EditoastPathOperationalPoint (which is the data from
  Editoast) and PathOperationalPoint (with a unique waypointId, and an
  optional opId)
- Switches to the relevant types in both STDCM and Operational Studies
- Stored the proper opId in the two places where the front generates
  unique IDs for the OPs for the Manchette
- Adapts upsertMapWaypointsInOperationalPoints so that it can handle
  both PathOperationalPoints and EditoastPathOperationalPoints

Signed-off-by: Alexis Jacomy <alexis.jacomy@gmail.com>