front: stop using Graou types for XML import
When processing XML files, we go from XML to ImportedTrainSchedule,
then to editoast's TrainSchedule. The roundtrip to
ImportedTrainSchedule is unnecessary, and forces us to unnecessarily
comply with the shape of the data returned by Graou, with strings
everywhere, unused fields and a forced CI code that we don't always
have.

This will also allow us to simplify the import logic later.

Signed-off-by: Simon Ser <contact@emersion.fr>