core: move ETCS stop and slowdown position calculation to ETCS module
In MaxSpeedEnvelope, SimStop disappears and SimStopInfo becomes SimStop, with the position becoming an Offset.
LoA and EoA calculations (as well as the usedCurveType logic) are moved to the ETCS Simulator.
Converted a couple of java tests to kotlin for interoperability problems due to SimStop now containing an Offset instead of a Double.

Signed-off-by: Erashin <alwenn.charpentier@gmail.com>