tests: remove paced train from test about reception on closed signal
The paced train was added into a relatively complex test, specifically
testing different configuration of reception on closed signal.
Involving paced train in the same test is mixing up the purpose of the
test, removing it (in favor of an entire dedicated different test in
the parent commit) is better and clearer.

Also add comments to the test to give more context.


Signed-off-by: Jean SIMARD <woshilapin@tuziwo.info>