ui-charts: fixes issue with repeated waypoints
This commit fixes #11868.

The ManchetteWithSpaceTimeChart was crashing when moving to the linear
scale, and when there are are least two waypoints exactly at the same
position.

This commit addresses that issue by fixing how spaceScales is computed
in useManchetteWithSpaceTimeChart, ensuring there are no SpaceScale with
a 0 coefficient or an infinite size.

Signed-off-by: Alexis Jacomy <alexis.jacomy@gmail.com>