ui-charts: improves PathLayer performances
This commit improves the PathLayer performances, by filtering the
segments and points to render on both the visible and the picking
layers.

It only keeps points that intersect with the stage, and segments which
timeframe intersect with the visible stage timeframe. It means we keep
"rendering" things over and under the canvas, but the gain seems quite
good already.

Signed-off-by: Alexis Jacomy <alexis.jacomy@gmail.com>