editoast: authz: add `model::Subject` and `s/mod subject/mod identity`
The `model::Subject` has nothing to do with the definitions in the
former `mod subject`.  So it would be weird to keep that module name
knowing that.  The definitions there are used in the `StorageDriver` and
in `editoast::client` to register, designate, and manage users and
groups.  Therefore `mod identity` seems like a good name since that's
what they actually represent.

Signed-off-by: Leo Valais <leo.valais97@gmail.com>