front: update formatMargin to always return something
The back-end accept to receive undefined margins but if it does, will return the value `{ boundaries: [], values: ['0%'] }`.
This can lead to mismatch when trying to compare two timetable items, one that has already been passed by the back end and one not.
Now if margins have only one value of 0%, return it instead of undefined.

Signed-off-by: SharglutDev <p.filimon75@gmail.com>