front: fill occurrence with its exception and refacto useOccurrences
If an occurrence is an exception, update it with the exception fields so the user can see the difference in the timetable UI.
As each occurrence displays its rolling stock image, we need to handle the rolling stock fetching directly in useOccurrences.
To handle this async call to get the rolling stock details, refacto useOccurrences from a useMemo to a useState/useEffect design.

Signed-off-by: SharglutDev <p.filimon75@gmail.com>