front: unify paced field in ngeToOsrd
There are two spots where we need to create the paced field: when
a new trainrun is created, and when a trainrun is updated. We used
to have two different codepaths for each. Refactor
createPacedAttributesFromTrainrun() so that it can be used from
handleUpdateTimetableItem().

The null check operator "!" makes it explicit in the code that we
assume a new trainrun will never have TRAIN_SCHEDULE_FREQUENCY_ID
(see the comment above handleCreateTimetableItem()).

Signed-off-by: Simon Ser <contact@emersion.fr>