core: add isClosedSignal flag to stops in maxSpeedEnveloppe inputs
We need to know whether we are stopping on an open or closed signal
in order to be able to use the proper etcs braking curve,
thus we add the isClosedSignal flag to the stops inputs passed
to maxSpeedEnveloppe.

This is just an enabler, there are no braking curve behavior changes
in this commit.

Signed-off-by: Alice Khoudli <alice.khoudli@polytechnique.org>