core, tests: use permitted speed with no svl on open signal in maxSpeedEnvelope etcs braking curve
A stop signal being open means that the danger zone following the signal is actually safe.
In etcs, we use a svl braking curve to protect this zone, which is thus not needed.

Thus we set the svl curve to null when computing our final etcs braking curve in maxSpeedEnvelope
if !isClosedSignal.

Additionally we only need to use the permitted speed curve instead of the more restrictive
indication speed curve when stopping on an open signal.

Overall we will brake later when stopping on an open signal in etcs, with no changes on closed signals.

We update the existing integration tests to use closed signals, to keep their behavior as is,
additional integration tests using open signals will be implemented later.

Signed-off-by: Alice Khoudli <alice.khoudli@polytechnique.org>