front: drop extractDateAndTime()
This function is a bit too specific to be lifted up to a file
containing generic date utils (it's used in a single spot).

Some functionality is unused: the date format is always set to
'DD/MM/YY', and the extracted hours/minutes are unused.

The fixed format string doesn't fare well in terms of localization:
someone from the US would expect the month to come before the day.

To fix these issues, drop the function and replace it with the
standard library.

Signed-off-by: Simon Ser <contact@emersion.fr>