tests: add test for etcs simulation with open signal stop
Add an integration test using a stop on an open signal to test
the behavior of the etcs braking curve.

It is a copy of an existing test which used a closed signal stop.
We check that the braking is performed later than in the closed signal case,
and that we follow the eoa permitted speed curve with no svl.

Signed-off-by: Alice Khoudli <alice.khoudli@polytechnique.org>