python, tests: fix pyright config not including each module files
Running pyright without file arguments and without proper includes lead to pyright ignoring most files,
which meant most files were not typechecked by the ci.

This lead to type errors in tests, fixed in previous commits.

In project dependencies are now specified as extrapath, not as include, which correspond to the files
to type check.

Signed-off-by: Alice K. <alice.khoudli@gmail.com>