editoast, front: add delete stdcm search env route
Currently stdcm search environments only support a create and retrieve latest route.
This mean thaat there is no way without editing the database to delete
a created stdcm search env.

In particular, this means that any env created by the e2e-tests can not be deleted
by the e2e-tests, and neither can the test infra, as the stdcm search env table uses
the infra id as a foreign key.

Thus this commit adds a route to delete the id to views/stdcm_search_environment,
adds a unit test for this route, and updates the openapi api.

Signed-off-by: Alice Khoudli <alice.khoudli@polytechnique.org>