editoast: add list route to stdcm environment
To be able to delete all stdcm search envs created by the e2e-tests,
including those created by previous e2e-tests that failed to execute
their teardown succesfully (either due to keyboard interrupt at any point
or an error during the teardown), we need to be able to list all envs
using the test infra.

The simplest, most reusable way to solve this requirement is to add a
generic list route.

Thus this commit adds a list route to the stdcmenv route,
a unit test for that route, and updates the openapi api.

Signed-off-by: Alice Khoudli <alice.khoudli@polytechnique.org>