front: fix destination stop duration in nge
In micro, to make the train stop at its destination, we add a stopFor 0 on the last pathStep.
In nge side, as the destination doesn't have any arrival or departure by default, we were erasing this schedule.
This was an issue if the paced train had been modified one time in nge, thus erasing the default schedule behavior, because to compute a paced train exception we compare the schedule of the occurrence being modified (which gets the stopFor 0 on last pathStep) and the one from the paced train.

Signed-off-by: SharglutDev <p.filimon75@gmail.com>