front: don't rely on TRAIN_SCHEDULE_FREQUENCY_ID in ngeToOsrd
This function will be used when importing foreign NGE files. In the
wild, NGE files will not use the same time frequencies as OSRD. In
particular, we can't rely on IDs used in the default OSRD time
categories.

Instead, only rely on the frequency value in minutes, and assume that
a frequency of 24 hours or more describes a single train.

Signed-off-by: Simon Ser <contact@emersion.fr>