front: avoid using an outdated updateTrainDepartureTime()
This callback depends on timetableItems, ie. it's re-generated
when the timetable item list changes. However, we were capturing
it from the broadcast channel useEffect() without a proper dep
on it.

Since we don't want to re-create the broadcast channel each time
the timetable item list changes, store the callback in a ref so
that we always use the latest one.

Signed-off-by: Simon Ser <contact@emersion.fr>