front: don't send HTTP request when receiving departure train broadcast
When dragging a train in the space time chart, we send an HTTP
request to editoast to update the database, then we update our
state, then we broadcast a message to all other windows. When
receiving the message, we send the same HTTP request and update
our local state.

Stop sending the HTTP request when receiving the broadcast message.

This has the nice side-effect of fixing an orthogonal issue:
updateTrainDepartureTime() depends on timetableItems, ie. it's
re-generated when the timetable item list changes. However, we were
capturing it from the broadcast channel useEffect() without a proper
dep on it. The new setTimetableItemDepartureTime() function doesn't
have this dependency so it fixes the issue.

Signed-off-by: Simon Ser <contact@emersion.fr>