front: surface uncaught JS errors to the user
Sometimes JavaScript errors are thrown in a callback (e.g. click
handler), or in asynchronous code. This can happen for a number
of reasons, for instance when accessing an array with a bad index.
React only catches errors in synchronous component logic, so
will miss these.

The error is logged to the browser console, but is otherwise
invisible to the user. This can be confusing: from the user's
point-of-view, clicking on a button will result in nothing
happening without any kind of feedback. Additionally, sometimes
these errors result in the app breaking in subtle ways (e.g.
the train list missing some updates, or the space time chart
missing some trains).

Surface the error to the user, so that they can open a bug report.
A bug report with a JS error details is way easier to debug,
especially for transient or hard-to-reproduce bugs.

To test, add the following to any component:

    useEffect(() => {
      setTimeout(() => {
        foo();
      }, 10);
    }, []);

Note, error handling in React will be improved once we switch to
version 19:
https://react.dev/blog/2024/12/05/react-19#error-handling

Signed-off-by: Simon Ser <contact@emersion.fr>