front: drop react-id-generator
react-id-generator has almost entirely been replaced with react's
built-in useId() hook or proper keys.

Use UUIDs for path step IDs. We were doing that in some places,
but were still using nextId() in some others, which can result in
collisions.

Use UUIDs for the local ID used for speed limit tags. The tag
itself can't be used, because the tag is mutable and we allow
intermediate states with duplicate tags. The UUIDs are generated
once for existing tags when the component is mounted, and once
for new tags when appending to the list, so they are stable.

Signed-off-by: Simon Ser <contact@emersion.fr>