front: correct css tabs for Horizon scrollbars
An old bug dating from December stated that two
scrollbars were appearing on the tabs component.
This was due to the fact that the tabs component
was not correctly overriding the default
Bootstrap SNCF styles.
Fixes are: _overrideBootstrapSNCF.scss got changed
so that the horizontal scrollbar is not displayed
anymore.
_tabs.scss got changed so that the text inside the
tabs could fit better on smaller screens (was
not linked with the bug).

missed a changed on div call dsg container

Signed-off-by: AlexandreRGJ <135651159+alexandrergj@users.noreply.github.com>