front: use currentData for simulationResults endpoints
Sometimes, console errors were printed when formatting the pathProperties due
to desynchronization between several endpoints (especially simulation and path
properties).
This bug (#11256) became more critical with my previous commit, as it now breaks
the application and results in blank pages.

To fix this issue, we now use the currentData from the endpoints. This ensures
that while the data is still being fetched, the component handles an undefined
value instead of the previously cached one.

Signed-off-by: Clara Ni <clara.ni@outlook.fr>