front: useSimulationResults now handles exceptions
The bug listed as 12281 is caused by useSimulationResults
not being able to handle exceptions. When adding
an exception, disabling the first occurence
triggered the error 'The occurence id should
match the format "indexedoccurrence_{pacedTrainId}_{occurrenceIndex}"'

A cheap an quick fix is to return undefined
when the selectedTrainId is an exception in the
useSimulationResults hook.

Signed-off-by: AlexandreRGJ <135651159+alexandrergj@users.noreply.github.com>