front: change stdcm contact mail to deployment setting
Change the contact us email adress in the stdcm help section
to the feedback mail defined in the deployment settings.

This feedback mail was previously only used in the contact form under the
simulations, but users would prefer the form from the help section to
also use this email, as the current static one was not the most relevant.

Signed-off-by: Alice K. <alice.khoudli@gmail.com>