front: memorize if occurence list open in train list
The state of pacedTrainItem is lost when the component is no
longer rendered, and we only render part of the list for performance
reasons. Thu whether the occurence list of a paced train is opened or not
is lost when scrolling to another part of the train list.

Thus keep the set of pacedTrainItem id with opened occurence list
in the timetable parent component containing the train list,
which is always rendered.

Signed-off-by: Alice K. <alice.khoudli@gmail.com>