front: rename scss white and black colors
Colours have a tint and multiple variants which are set to ensure a certain contrast level in order to respect WCAA guidelines.
Ex: grey30, primary70, or waring60.
Black and White variants are obtained by reducing their opacity. Black50 is a black with 50% opactity.
The number suffix is representing a contrast level for some colours, and an opacity for black and white, which is confusing and may result in interpretation errors.
The correction aims at being more precise and add in the colour name the opacity effect it carries. Ex: BlackAlpha25.

Signed-off-by: Benoit Simard <contact@bsimard.com>