front: display tdcm perimeter on map
Adding perimeter in redux state and load it from the API.
Due to the debug mode on which we need to see the full data, I reset the
perimeter on the useEffect clean up function.

Changing map color globally according to the design, and adding a
muted color

On each infra layer, adding the attribut `higlightedArea` to be able
to filter or design the data. Every data outside of the area is
filtered, except for tracks and POI on which data are muted.

On LMR, zooming on the highlighted area per default (see StdcmMapActivePerimeter)

Signed-off-by: Benoit Simard <contact@bsimard.com>