front: use matchPathStepAndOp in times stops output table
In operational studies simulation results,
the times stops output table was matching path steps and op purely based on positionOnPath,
while the simulation sheet export was using matchPathStepAndOp like the rest of the application.

Additionally, the pure position check was not working with https://github.com/OpenRailAssociation/osrd/pull/10200
due to path steps position being offset from the op on timetable created to serve as stdcm envs.

Thus unifies the times stops output table to use matchPathStepAndOp like other components.

It is worth mentioning that we will soon want to introduce a position check in matchPathStepAndOp,
which will be an issue due to #10200. However, this would have been an issue anyway for the other components using it.
Using matchPathStepAndOp in the output table mean we only need to address the issue in matchPathStepAndOp,
rather than in each individual component.

Signed-off-by: Alice Khoudli <alice.khoudli@polytechnique.org>