editoast: adding stdcm perimeter
Adding column in table. Its type is `json` and not Geometry because
the "geo model" is not (yet?) developped, and model is needed in the
API (using changeset, pagination).
Because this data is used "as it" on the UI  (ie. no geo computation in postgis), it's OK for now to store it as JSON.
And it will be easier in the futur to upgrade to the "geo model" (the
other solution was to rewrite specific function for this model).

CLI has been also updated, so you can specify a path to a geojson file
that containbs the geometry.

Signed-off-by: Benoit Simard <contact@bsimard.com>