front: fix parsing of datetime-local input
Fix #12502

In dayjs, parsing a date with 2 digits (or less) - which is the case when you start typing something - results in 1900+ "what you started to type".
Check https://github.com/iamkun/dayjs/issues/1237

So removing the use of dayjs in the `parseLocalDateTime` function which
is only used for the datetime-local input.

Signed-off-by: Benoit Simard <contact@bsimard.com>