core, editoast, front: add stops_at_end_of_block option to TrainSchedule
This commit uses `stops_at_end_of_block` flag to deal with the issue that the trains were previously stopping exactly on the Operational Point Part position, which is often in the middle of the train platform. The position of the train is considered at the head of the train. This is an issue, because in the real life, a train (meaning its head) stops at the end of the platform (at a specific stop panel position), allowing the rest of the train not to overflow of the passed tracks, avoiding unnecessary track occupation. Also, this doesn't take account of the train direction.

This implementation makes the pathfinding algorithm move the stops of these trains closer to the end of the current block, to try to avoid that the tail of the train is on another track, and also never further than the next block-delimiting signal.

- Operational Studies:
	- Nothing changed except option propagation: `stops_at_end_of_block = false` (default).
- STDCM:
	- The trains that compose the timetable (search environment) are all simulated with `stops_at_end_of_block = true`.
	- The simulated train is simulated also with this flag `stops_at_end_of_block = true`.

Note: this should be reverted when we have the stop panels.

Signed-off-by: Louis Greiner <greiner.louis@gmail.com>
Signed-off-by: Angelina Kuntz <angelina.kuntz@epita.fr>