ui-charts: use internalHeight instead of height in SpeedSpaceChart
Previously, the chart managed its own height, which could vary either when the user enlarged it or when they added/
removed linear layers (dynamicalHeight was private to the chart).

We now need to control the entire height of the graph from the parent component, hence the introduction of internalHeight,
which corresponds to the height of the graph (without the linear layers and with its top and bottom margins) and is used
in the various layers to replace height and dynamicalHeight.

Co-authored-by: Louis <iLTmpas@users.noreply.github.com>
Signed-off-by: Clara Ni <clara.ni@outlook.fr>