editoast: model: fix orphan rule for Model fields associated items
The hack consists of removing the `ModelField` wrapper to replace it by
a specific wrapper defined next to the model type.  It's far from ideal,
but the original implementation leaves much to be desired as well.
Although it's not worth blocking the crates split for that, the whole
listing paradigm should be revised at some point.

Signed-off-by: Leo Valais <leo.valais97@gmail.com>