front: ngetoosrd: always consider nge stops as stops
When lacking a departure time, nge stops (node transitions marked with isNoneStopTransit set to false)
were not considered as stops by osrd, and the stop_for property of the corresponding
path steps was set to null.

This could lead to a return trip not having the same stops as the forward trip,
leading to an incompatibility.

Signed-off-by: Louis Greiner <greiner.louis@gmail.com>
Co-authored-by: Alice Khoudli <alice.khoudli@polytechnique.org>
Co-authored-by: Simon Ser <contact@emersion.fr>