editoast: core_client: fix initial connection always triggering a retry
When creating a new client, we start the connection loop, itself calling
connection_ok.  The latter fails the first time since there's no
connection open, triggering the reconnection routing to acquire the
first connection.  However, in the meantime, the `new` function also
probes `connection_ok`, which isn't ready yet, so we retry after one
second, delaying the readiness of editoast.

Here we just send a signal the first time the connection has been
acquired so we don't waste the first retry and pay for the delay.

While experimenting, I've implemented the exponential delay logic, which
can't hurt so I left it ¯\_(ツ)_/¯

Signed-off-by: Leo Valais <leo.valais97@gmail.com>