front: add reducers to unset selected projected ids
Add reducers to check whether the selected and projected train ids in the store
match with a given scheduled or paced train id,
an occurence of a given paced train id,
or an occurence of a given paced train id not listed in a list of present ids.

These will be used to properly unset selected and projected train ids when
deleting timetable items or occurences (possibly simply by changing paced train pacing parameters)
without needing all components that perform such operations to consume the selected and projected train ids.

Ids were currently only unset in a few places but not others,
causing both app crashes and unintended user experience.

Signed-off-by: Alice K. <alice.khoudli@gmail.com>