front: revert TrainMainCategoryDict to record
TrainMainCategoryDict was created as a passthrough record over an enum
or a list for the sole purpose of enforcing at compile time that the list
of categories used in the front is the exact same as the one
provided by the open api type (which no longer exists at runtime).

Using a map does not provide this same type safety.
Additionally, this is a passthrough dict, there is no reason to
use get on it.

Signed-off-by: Alice Khoudli <alice.khoudli@polytechnique.org>