front: fix electrification ranges and profiles merging
The function formatElectrificationRanges aimed to merge together electrification ranges
and electrical profile ranges.

However, this was done assuming that the boundaries of electrical profiles would be the
exact same or included in the boundaries of electrification ranges, which is not true.

The electrical profiles data produced by formatElectrificationRanges was thus no longer used
in the simulation csvs, as it was incomplete.

In order to use them again, properly merge these ranges together.

The easiest way to properly do so is actually to merge them before converting from boundaries to ranges.

Thus invert the order of these operations.

Signed-off-by: Alice Khoudli <alice.khoudli@polytechnique.org>