front: fix round trip modal crashing on older browsers
Reduce over iterators is a newer feature, that was recently implemented in 2025 in firefox.
Older browsers, including the older extended support release of firefox, do not support it.
Since it's not really needed (there are many other ways to loop!),
it would be best to avoid using it, at least for a couple years,
as it would otherwise impact many of our users.

Signed-off-by: Alice Khoudli <alice.khoudli@polytechnique.org>