front: fix start time for return trip in NGE
We were picking the start time of the forward train for both trains
of the round-trip.

The return start time needs to be computed separately. Move this
logic out of generateTrainrunProperties(), so that this function
only computes properties common to both trains.
generatePathAndSchedule() is now responsible for computing and
formatting the start time.

Instead of reversing train run sections before calling
calculateStartDate(), pick the last element when appropriate.

Signed-off-by: Simon Ser <contact@emersion.fr>
Closes: https://github.com/OpenRailAssociation/osrd/issues/13030