ui-charts: prevents nil hoveredItem cases
This commit fixes #13408.

The best solution would be to find and solve the cause for the cases
where pickingElements is empty in useCanvas.

Because I did not find this cause indeed, this commit simply mitigates
this issue, by ignoring the cases where it happens, and commenting the
code for future brave maintainers.

Signed-off-by: Alexis Jacomy <alexis.jacomy@gmail.com>