front: fix exceptions tooltip positioning
The tooltip was rendered below the viewport, as its top value was too high.
Indeed, getBoundingClientRect returned a top position relative to the viewport,
but fixed positioning seems to be affected by containing blocks,
and not always be directly fixed relative to the viewport.
Here, one parent of each exception's occurrence-indicator seems to have contain:strict.

To fix this, renders the tooltip outside of the hijacking containing block
and directly in the top layer over everything using a popover.

Resetting inset is necessary here to allow bottom positioning to function and avoid default top value.

zIndex is no longer needed thanks to popover.

Finally drop isOpen as we always only render the tooltip conditionally anyway, and this simplifies using popover.

As an aside, as soon as anchor positioning is supported in firefox, we should switch to it.

Signed-off-by: Alice Khoudli <alice.khoudli@polytechnique.org>