front: simplify conflicts fetching
Conflicts were manually fetched using lazy query,
with manual cache invalidation whenever a function updating the timetable was called.
A separate useEffect handled the initial fetch on page load.
This strategy worked, but was complex and place burden on developers to remember
to add code to each function that could affect the timetable.

However, we could simply use useQuery and let rtk query handle the cache invalidation,
as the query is already dependent on the 'timetable' tag.
This would be both simpler and less brittle.

The change from useQuery to useLazyQuery was previously made due to issues with manual refetching when using useQuery.
I do not believe these manual refetches were actually necessary (as it is already handled by tags),
and I believe we can thus revert to useQuery while dropping the refetching logic.

Signed-off-by: Alice Khoudli <alice.khoudli@polytechnique.org>