editoast, front: inline non-built-in schemas in path & query params
The upcoming utoipa schema collection mechanism doesn't recurse in
`IntoParams` definitions. Which means we either have to collect such
schemas manually, or inline them. But on the flip side, we won't have to
manually mark which schemas to collect ever again.

Some of the schemas that got inlined are used by the frontend. The types
are added back in the interface.

Signed-off-by: Leo Valais <leo.valais97@gmail.com>