editoast: fix missing fields in scenario and study search results
Paced trains count was used by the front in addition to trains count when using the search endpoint,
and was present in ScenarioResponse but not SearchResultItemScenario.

Study type, business code and study code were used by the front when using the search endpoint,
and were present in Study but not SearchResultItemStudy.

Signed-off-by: Alice Khoudli <alice.khoudli@polytechnique.org>