docker: wait for dev front container to be healthy
The development frontend container might take a while to start,
because it installs dependencies and builds osrd-ui.

This might confuse users because the gateway will throw an HTTP
error while the front container is starting up. Make it clear that
the development setup is not fully started by waiting for the
front container to be healthy.

Bump the start_period because 4s is definitely not enough for the
front container to download all dependencies and build osrd-ui.

Signed-off-by: Simon Ser <contact@emersion.fr>