front: upgrade eslint-plugin-react-hooks to v7
The new version introduces a whole bunch of new rules. These new
rules trigger hundreds of errors across the project.

For now, disable rules that cause errors so that we can upgrade
the dependency. We'll be able to fix these errors incrementally.
Upgrading is still worthwhile to ensure we don't introduce more of
these and don't fall further behind on NPM updates.

Signed-off-by: Simon Ser <contact@emersion.fr>