front, ui-icons: fix ESLint ignore rule
ui-icons has a bunch of auto-generated files we don't want to run
ESLint on. However, since the ESLint flat config migration we're
analyzing these files and generating a long list of errors.

ESLint assumes that a lone { ignores } is a global ignore
configuration object:
https://eslint.org/docs/latest/use/configure/configuration-files#globally-ignoring-files-with-ignores

Having a "files" field in the same object makes this a local
configuration object. To fix this, split it in half.

To test:

    cd front/ui/ui-icons
    npm run build

This should not print any error.

Signed-off-by: Simon Ser <contact@emersion.fr>