front: fix track offset nodes in NGE
When building an OSRD train path, we were unconditionally using
the trigram from the NGE node's betriebspunktName field. This worked
fine for OPs with a trigram, but fell apart for track offsets (and
infrastructures without trigrams).

Instead, reconstruct the PathItemLocation from the path_item_key
stored in MacroEditorState.nodes.

As a bonus, this makes it so path item location types are no
longer changed to trigrams when editing a train from NGE: if a
train was created with UICs from OSRD, editing it in NGE will
retain UICs now.

We still need to keep the old betriebspunktName-based logic for
NGE JSON file imports.

Signed-off-by: Simon Ser <contact@emersion.fr>
Closes: https://github.com/OpenRailAssociation/osrd/issues/12618