front: fix missing hook dependencies in useOutputTableData()
This useMemo() reads scheduleByAt without having it specified in
its dependencies.

Specifying scheduleByAt in dependencies wouldn't work: this variable
is not memoized and is re-computed at each render, so the useMemo()
would be triggered each render too.

Instead, move the scheduleByAt declaration inside the useMemo()
callback, and declare that the useMemo() depends on selectedTrain.

Same applies to theoreticalMargins.

Signed-off-by: Simon Ser <contact@emersion.fr>
Closes: https://github.com/OpenRailAssociation/osrd/issues/13767