editoast: telemetry: allow to disable the batch export of spans
Make `common::tracing::create_tracing_subscriber` take an additional
parameter that allows to setup tracing to either send spans by batch
(as it was always the case previously) or to send them individually in a
blocking way.

Disabling batch processing of traces will prevent losing some of them
when using `common::tracing` to setup the telemetry for small
executables with a limited lifespan (example: executing a CLI binary).

Signed-off-by: Loup Federico <16464925+Sh099078@users.noreply.github.com>