editoast: fix invalid make-migration generated diesel migration revert
The `make-migration` cli command generates up and down sql migration
scripts that delete their related triggers and functions. It causes
compatibility issues if prior migrations manipulate the same objects.

Update the generated `2023-12-19-122428_search_op_ci_code/down.sql`
script so that it correctly reverts to the database schema prior state
instead of deleting its related PostgreSQL objects.

This commit is a quick fix on the faulty down.sql generated script. It is
a fast way to make migration reverts work again but it doesn't fix the
root cause of the issue: `make-migration` should be updated so that we can
revert the PostgreSQL objects it manipulates instead of deleting them in
the generated down.sql scripts.

Signed-off-by: Loup Federico <16464925+Sh099078@users.noreply.github.com>